alert timeout function  is set in admin/inc/scripts.php

+++ admin/ajax/feature_facilities.php
@@ -12,8 +12,6 @@ if(isset($_POST['add_feature']))
  $values = [$frm_data['name']];
  $res = insert($q, $values,'s');
  echo $res;
-
-
 }
 
 if(isset($_POST['get_features']))
@@ -47,6 +45,8 @@ if(isset($_POST['rem_feature']))
    $res = delete($q,$values,'i');
    echo $res;
 }
+
+// functions for facilities
 
 if(isset($_POST['add_facility']))
 {
@@ -97,12 +97,22 @@ if(isset($_POST['get_facilities']))
 
 if(isset($_POST['rem_facility']))
 {
-   $frm_data = filteration($_POST);
+   $frm_data = filteration($_POST);   
    $values = [$frm_data['rem_facility']];
 
-   $q="DELETE FROM `facilities` WHERE `id`=?";
-   $res = delete($q,$values,'i');
+   $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+   $res= select($pre_q,$values,'i');
+   $img = mysqli_fetch_assoc($res);
+
+   if(deleteImage($img['icon'],FACILITIES_FOLDER)){  //this block of code deletes icon from folder and db 
+    $q="DELETE FROM `facilities` WHERE `id`=?";
+    $res = delete($q,$values,'i');
    echo $res;
+   }
+   else{
+     echo 0; 
+   }
+   
 }
 
 ?>
